requirements.txt
flask==3.1.0
flask-sqlalchemy==3.1.1
flask-login==0.6.3
flask-cors==5.0.1
flask-migrate==4.0.7
sqlalchemy==2.0.36
werkzeug==3.1.3
python-dotenv==1.0.1
anthropic==0.42.0
requests==2.32.3
gunicorn==23.0.0
PyPDF2==3.0.1
python-docx==1.1.2
psycopg2-binary==2.9.10
sentry-sdk[flask]==2.0.0
